# Conflicts:
#	DynamicSurvey.Server/Controllers/SurveysController.cs
#	DynamicSurvey.Server/DynamicSurvey.Server.csproj
#	packages/Antlr.3.5.0.2/Antlr.3.5.0.2.nuspec

+++ DynamicSurvey.Server.DAL/Repositories/UsersRepository.cs
@@ -90,7 +90,7 @@ namespace DynamicSurvey.Server.DAL
 				AccessRight = new AccessRight()
 				{
 					Name = Convert.ToString(row["UserRight"]),
-					AccessLevel = (AccessLevel)(row["AccessLevel"]),
+					AccessLevel = (AccessLevel)(Int32)(row["AccessLevel"]),
 					Id = (ulong)(row["UserRightId"])
 				}
